Tests: Make getHandleCenterPoint a shared util.
